\name{hapmapsnp5-package}
\alias{hapmapsnp5-package}
\alias{hapmapsnp5}
\docType{package}
\title{
  Sample HapMap SNP 5.0
}
\description{
  
  Sample dataset on the SNP 5.0 Affymetrix platform. Data obtained from
http://www.hapmap.org . The package is meant to be used only for
demonstration of BioConductor tools.
  
}

\details{
\tabular{ll}{
Package: \tab hapmapsnp5\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab March/2007\cr
License: \tab GPL\cr
}

Details on the data are available at http://www.hapmap.org .

The maintainer does not warrant the accuracy of the data.

}
\author{
  Data generated by the HapMap Consortium.

  Package maintained by Benilton Carvalho <bcarvalh@jhsph.edu>.
}
\references{
The International HapMap Consortium. The International HapMap
Project. Nature 426, 789-796 (2003).
}

\keyword{ package }
\examples{
## library(oligo)
## library(hapmapsnp5)
## the.path <- system.file("celFiles", package="hapmapsnp5")
## cels <- list.celfiles(path=the.path, full.names=TRUE)
## temporaryDir <- tempdir()
## rawData <- read.celfiles(fullFilenames, tmpdir=temporaryDir)
}
